+++ packages/apollos-ui-fragments/src/index.js
@@ -1,17 +1,17 @@
 import * as content from './content';
 import * as features from './features';
 import * as live from './live';
-import * as profile from './profile';
-import * as scripture from './scripture';
 import * as pass from './pass';
+import * as scripture from './scripture';
+import * as user from './user';
 
 const FRAGMENTS = {
   ...content,
   ...features,
   ...live,
-  ...profile,
   ...scripture,
   ...pass,
+  ...user,
 };
 
 export default FRAGMENTS;
